o  fixed int references in rpc client.
git-svn-id: https://svn.apache.org/repos/asf/maven/continuum/branches/id-refactor@497960 13f79535-47bb-0310-9956-ffa450edef68

+++ continuum-rpc-client/src/main/java/org/apache/maven/continuum/rpc/ProjectsReader.java
@@ -92,7 +92,7 @@ public class ProjectsReader
     {
         XmlRpcClient client = new XmlRpcClient( server );
         Vector vect = new Vector();
-        vect.add( new Integer( proj.getId() ) );
+        vect.add( new Long( proj.getId() ) );
         Object obj = client.execute( "continuum.getProject", vect );
         if ( obj instanceof Hashtable )
         {
@@ -110,7 +110,7 @@ public class ProjectsReader
     {
         XmlRpcClient client = new XmlRpcClient( server );
         Vector vect = new Vector();
-        vect.add( new Integer( proj.getId() ) );
+        vect.add( new Long( proj.getId() ) );
         //trigger
         vect.add( new Integer( 1 ) );
         Object obj = client.execute( "continuum.buildProject", vect );
